requirements.txt
PyYAML
attrs==19.3.0
Automat==0.8.0
constantly==15.1.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
paho-mqtt==1.5.0
PyHamcrest==1.9.0
pymodbus==2.3.0
pyModbusTCP==0.1.8
pyserial==3.4
six==1.13.0
Twisted==19.10.0
zope.interface==4.7.1
